Bash - safely handle services output file create and delete

+++ bash/get_services.sh
@@ -3,8 +3,10 @@
 # Refs:
 # https://askubuntu.com/questions/407075/how-to-read-service-status-all-results
 
-rm ./output/services-services.txt
-touch ./output/services-services.txt
+if [ -f "./output/services-services.txt" ]; then 
+    rm ./output/services-services.txt
+    touch ./output/services-services.txt
+fi
 
 sudo service --status-all >>./output/services-services.txt 2>&1
 sudo initctl list > ./output/upstart-services.txt
